fastapi-app/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.29
sqlalchemy>=2.0,<3.0
alembic>=1.13
pydantic>=2.0
bcrypt>=4.0
python-dotenv
jinja2
python-multipart
itsdangerous
